+++ docs/changelog.rst
@@ -8,8 +8,10 @@ Release Notes
 Bug fixes
 ~~~~~~~~~
 
-* Use the codeobj name in the key rather than the line number.
-  (Reported by David Charboneau in `Issue 174`_ and fixed by David Charboneau in `PR 175`_.)
+* Use ``codeobj.__name__`` in the key for the internal cache of
+  :class:`astor.file_util.CodeToAst` rather than the line number to
+  prevent :exc:`KeyError`.
+  (Reported and fixed by David Charboneau in `Issue 174`_ and `PR 175`_.)
 
 .. _`Issue 174`: https://github.com/berkerpeksag/astor/pull/174
 .. _`PR 175`: https://github.com/berkerpeksag/astor/pull/175